spi: cadence-quadspi: Use STIG for small reads
c9293d064e11 (spi: cadence-quadspi: do not use DMA for reads smaller
than 16 bytes, 2021-04-29) started using CPU based copy for small reads
to avoid the DMA overhead. This is completely correct in principle. But
for inexplicable reasons, the next register write goes missing if this
is used on TI's J7200 EVM. This register write is for setting the number
of dummy cycles when switching to Octal DTR mode. Due to this missing
register write subsequent reads are missing the first 8 bytes. What is
even more surprising is that this problem is not seen on AM64x-EVM or
AM64x-SK, which have the exact same flash and controller combo.

Avoid that code path completely, and use the STIG controller for small
reads instead. This is the better idea to begin with anyway since the
STIG controller is designed to access flash registers, which are the
main source of small reads.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>