media: ti: j721e-csi2rx: reduce max frame height and width
Currently the max width and height are set to UINT_MAX. This has
multiple problems. First, max_width must be a multiple of step_width,
which is not accounted for. Round down max_width to a multiple of
step_width.

Second, Gstreamer treats the numbers as signed values, and trips up when
dealing with such large values since they are interpreted as negative
values. While it is not generally right to change the kernel for an
application bug, it is not such a big deal here. This change is
replacing one set of absurdly large arbitrary values with another set of
smaller but still absurdly large arbitrary values. Both limits are
unlikely to be hit in practice.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>