media: ti: j721e-csi2rx: Drop pipeline validation
The pipeline validation being done here is completely broken. Firstly,
it assumes that the media_graph_walk_* APIs perform a depth-first
traversal on the media graph. They claim to do so, but that is not the
case in reality. The order of traversal seems to not be guaranteed, just
that all the entities in the graph will be traversed at some point. This
means that the device that last touched for format can't be reliably
found.

Secondly, a driver is not supposed to validate format propagation
through the entire pipeline anyway. It should just mind its own formats
and its own links, and should not traverse the graph to validate the
entire thing.

Due to these two reasons, it is better to just drop the validation.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>