spi: cadence-quadspi: do not use DMA for reads smaller than 16 bytes
For reads this small it does not make a lot of sense to go through DMA
which usually has a significant setup overhead. The benefits of DMA are
reaped in larger transfers. This is especially relevant in case a
register is being polled. The DMA overhead will make the polling slower
than it should be.

This commit was reverted in commit 38a51ccb9448 ("Revert "spi:
cadence-quadspi: do not use DMA for reads smaller than 16 bytes"")
because it caused some register writes to go missing on J7200. As a
result some flash configuration did not go through and the flash stopped
working properly. This has since been fixed so this commit can be
reinstated.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>