arm64: dts: ti: Add AM62x SK camera overlay
Add an overlay for OV5640 camera sensor for AM62x SK. The camera module
used is Digilent PCam 5C, which provides OV5640 sensor with the
Raspberry Pi connector.

The module does not work at 400 KHz reliably. Drop the I2C bus frequency
to 10 KHz to make it work properly. The only other peripheral on the bus
is the expansion header so the impact should not be too large.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>